requirements.txt
streamlit
pyiceberg
httpx
kafka-python
mapr-streams-python
